requirements.txt
google-adk==1.14.1
psutil
litellm==1.77.1
google-generativeai==0.8.5
python-dotenv==1.1.1